--- src/main/kotlin/ru/hse/Executable.kt
@@ -1,8 +1,0 @@
-package ru.hse
-
-import java.io.InputStream
-import java.io.OutputStream
-
-interface Executable {
-    fun run(input: InputStream, output: OutputStream, error: OutputStream): ExecutionResult
-}
