--- src/main/kotlin/ru/hse/ExecutionResult.kt
@@ -1,3 +1,0 @@
-package ru.hse
-
-data class ExecutionResult(val exitCode: Int, val needExit: Boolean)
