--- src/main/kotlin/ru/hse/ExpressionExecutor.kt
@@ -1,8 +1,0 @@
-package ru.hse
-
-import java.io.InputStream
-import java.io.OutputStream
-
-interface ExpressionExecutor {
-    fun execute(expression: String, input: InputStream, output: OutputStream, error: OutputStream): ExecutionResult
-}
