--- src/main/kotlin/ru/hse/Parser.kt
@@ -1,5 +1,0 @@
-package ru.hse
-
-interface Parser<T> {
-    fun parse(line: String): Result<T>
-}
