--- src/main/kotlin/ru/hse/PipeFactory.kt
@@ -1,5 +1,0 @@
-package ru.hse
-
-interface PipeFactory {
-    fun create(commands: List<Executable>): Executable
-}
