--- src/main/kotlin/ru/hse/PipeParser.kt
@@ -1,7 +1,0 @@
-package ru.hse
-
-class PipeParser : Parser<List<List<String>>> {
-    override fun parse(line: String): Result<List<List<String>>> {
-        TODO("Not yet implemented")
-    }
-}
