--- src/main/kotlin/ru/hse/PipeSplitter.kt
@@ -1,5 +1,0 @@
-package ru.hse
-
-interface PipeSplitter {
-    fun split(tokens: List<String>): Result<List<List<String>>>
-}
