--- src/main/kotlin/ru/hse/Substitutor.kt
@@ -1,5 +1,0 @@
-package ru.hse
-
-interface Substitutor {
-    fun substitute(token: String): String
-}
