--- src/main/kotlin/ru/hse/Tokenizer.kt
@@ -1,5 +1,0 @@
-package ru.hse
-
-interface Tokenizer {
-    fun tokenize(expression: String): Result<List<String>>
-}
