--- src/main/kotlin/ru/hse/VarNameValidator.kt
@@ -1,6 +1,0 @@
-package ru.hse
-
-interface VarNameValidator {
-    fun check(token: String): Boolean
-    fun nameFromBeginningIn(token: String): String
-}
