+++ src/main/kotlin/ru/hse/cli/CLI.kt
@@ -1,0 +1,12 @@
+package ru.hse.cli
+
+import java.io.InputStream
+import java.io.OutputStream
+
+interface CLI {
+    fun getLine(): String
+    fun showMessage(message: String)
+    fun getInputStream(): InputStream
+    fun getOutputStream(): OutputStream
+    fun getErrorStream(): OutputStream
+}
