+++ src/main/kotlin/ru/hse/executable/Executable.kt
@@ -1,0 +1,8 @@
+package ru.hse.executable
+
+import java.io.InputStream
+import java.io.OutputStream
+
+interface Executable {
+    fun run(input: InputStream, output: OutputStream, error: OutputStream): ExecutionResult
+}
