+++ src/main/kotlin/ru/hse/executable/ExecutionResult.kt
@@ -1,0 +1,3 @@
+package ru.hse.executable
+
+data class ExecutionResult(val exitCode: Int, val needExit: Boolean)
