+++ src/main/kotlin/ru/hse/factory/CommandFactory.kt
@@ -1,0 +1,9 @@
+package ru.hse.factory
+
+import ru.hse.executable.Executable
+import java.util.function.Function
+
+interface CommandFactory {
+    fun registerCommand(name: String, factory: Function<List<String>, Executable>)
+    fun create(tokens: List<String>): Executable
+}
