+++ src/main/kotlin/ru/hse/factory/PipeFactory.kt
@@ -1,0 +1,7 @@
+package ru.hse.factory
+
+import ru.hse.executable.Executable
+
+interface PipeFactory {
+    fun create(commands: List<Executable>): Executable
+}
