+++ src/main/kotlin/ru/hse/parser/Parser.kt
@@ -1,0 +1,5 @@
+package ru.hse.parser
+
+interface Parser<T> {
+    fun parse(line: String): Result<T>
+}
