+++ src/main/kotlin/ru/hse/parser/PipeParser.kt
@@ -1,0 +1,7 @@
+package ru.hse.parser
+
+class PipeParser : Parser<List<List<String>>> {
+    override fun parse(line: String): Result<List<List<String>>> {
+        TODO("Not yet implemented")
+    }
+}
