+++ src/main/kotlin/ru/hse/splitter/PipeSplitter.kt
@@ -1,0 +1,5 @@
+package ru.hse.splitter
+
+interface PipeSplitter {
+    fun split(tokens: List<String>): Result<List<List<String>>>
+}
