+++ src/main/kotlin/ru/hse/substitutor/Substitutor.kt
@@ -1,0 +1,5 @@
+package ru.hse.substitutor
+
+interface Substitutor {
+    fun substitute(token: String): String
+}
