+++ src/main/kotlin/ru/hse/tokenizer/Tokenizer.kt
@@ -1,0 +1,5 @@
+package ru.hse.tokenizer
+
+interface Tokenizer {
+    fun tokenize(expression: String): Result<List<String>>
+}
