+++ src/main/kotlin/ru/hse/validator/VarNameValidator.kt
@@ -1,0 +1,6 @@
+package ru.hse.validator
+
+interface VarNameValidator {
+    fun check(token: String): Boolean
+    fun nameFromBeginningIn(token: String): String
+}
