+++ src/test/kotlin/ru/hse/command/ExitCommandTest.kt
@@ -3,7 +3,7 @@ package ru.hse.command
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.Arguments
 import org.junit.jupiter.params.provider.MethodSource
-import ru.hse.Executable
+import ru.hse.executable.Executable
 import java.io.ByteArrayInputStream
 import java.io.ByteArrayOutputStream
 import java.nio.charset.StandardCharsets
